[NC]: typing fix
